+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.8.0 2026-05-12
 ### Breaking change
-- Changed charger starting current from 6A to charger min current. Please ensure your charge min current is set.
+- Changed charger starting current from 6A to min workable current. Please ensure your min workable current is set.
 - Added "Min workable current enter pause percent" local device config.
 - Moved "Min workable current exit pause percent" from global defaults to local device config.
 - Added "Max consumed energy limit" and "End on max consumed energy" switch to end charge on reaching max consumed energy.